@@ -16,15 +16,18 @@
         "forum": "https://discourse.laminas.dev"
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     },
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0"
+        "php": "~8.0.0 || ~8.1.0 || ~8.2.0"
     },
     "require-dev": {
-        "laminas/laminas-coding-standard": "~2.2.1",
+        "laminas/laminas-coding-standard": "~2.4.0",
         "laminas/laminas-stdlib": "^2.7.7 || ^3.1",
-        "phpunit/phpunit": "^9.3"
+        "phpunit/phpunit": "^9.5.25"
     },
     "suggest": {
         "laminas/laminas-json-server": "For implementing JSON-RPC servers",
